+++ CSharpGL/Scene/Lights/LightBase.cs
@@ -1,105 +1,64 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel;
+using System.Drawing.Design;
 using System.Linq;
 using System.Text;
 
 namespace CSharpGL
 {
     /// <summary>
-    /// 
+    /// Base type of directional light, point light and spot light.
     /// </summary>
-    public abstract class LightBase : ITextureSource
+    [Editor(typeof(PropertyGridEditor), typeof(UITypeEditor))]
+    public abstract class LightBase
     {
-        private vec3 color = new vec3(1, 1, 1);
+        //private vec3 color = new vec3(1, 1, 1);
+        ///// <summary>
+        ///// Light's color.
+        ///// </summary>
+        //public vec3 Color { get { return this.color; } set { this.color = value; } }
+
+        private vec3 diffuse = new vec3(1, 1, 1);
         /// <summary>
-        /// 
+        /// Diffuse intensity.
         /// </summary>
-        public vec3 Color { get { return this.color; } set { this.color = value; } }
+        public vec3 Diffuse { get { return this.diffuse; } set { this.diffuse = value; } }
+
+        private vec3 specular = new vec3(1, 1, 1);
+        /// <summary>
+        /// Specular intensity.
+        /// </summary>
+        public vec3 Specular { get { return this.specular; } set { this.specular = value; } }
 
         /// <summary>
-        /// 
-        /// </summary>
-        public float AmbientIntensity { get; set; }
-
-        /// <summary>
-        /// 
-        /// </summary>
-        public vec3 DiffuseIntensity { get; set; }
-
-        /// <summary>
-        /// 
-        /// </summary>
-        public vec3 SpecularIntensity { get; set; }
-
-        /// <summary>
-        /// 
+        /// Light's position.
         /// </summary>
         public vec3 Position { get; set; }
 
-        private readonly IFramebufferProvider framebufferProvider = new DepthFramebufferProvider();
-        private readonly PolygonOffsetFillState state = new PolygonOffsetFillState(false);// TODO: other offsets also needed?
-        private readonly int[] viewport = new int[4];
+        /// <summary>
+        /// 
+        /// </summary>
+        [Editor(typeof(PropertyGridEditor), typeof(UITypeEditor))]
+        public Attenuation Attenuation { get; set; }
 
         /// <summary>
-        /// bind framebuffer, setup viewport, polygon-offset and so on.
+        /// Base type of all lights.
         /// </summary>
-        public void Begin()
+        /// <param name="attenuation"></param>
+        public LightBase(Attenuation attenuation)
         {
-            GL.Instance.GetIntegerv((uint)GetTarget.Viewport, viewport);
-            int width = viewport[2], height = viewport[3];
-            var framebuffer = this.framebufferProvider.GetFramebuffer(width, height);
-
-            framebuffer.Bind();
-            GL.Instance.Viewport(0, 0, width, height);
-            this.state.On();
-            {
-                GL.Instance.ClearColor(1.0f, 1.0f, 1.0f, 1.0f);// white color means farest position.
-                GL.Instance.Clear(GL.GL_COLOR_BUFFER_BIT | GL.GL_DEPTH_BUFFER_BIT | GL.GL_STENCIL_BUFFER_BIT);
-                //{
-                //    var args = new RenderEventArgs(this.Camera);
-                //    foreach (var item in this.Children)
-                //    {
-                //        RenderAction.Render(item, args, true, true);
-                //    }
-                //}
-            }
-        }
-
-        /// <summary>
-        /// unbind framebuffer, reset viewport, polygon-offset and so on.
-        /// </summary>
-        public void End()
-        {
-            int width = viewport[2], height = viewport[3];
-            var framebuffer = this.framebufferProvider.GetFramebuffer(width, height);
-
-            this.state.Off();
-            GL.Instance.Viewport(viewport[0], viewport[1], viewport[2], viewport[3]);// recover viewport.
-            framebuffer.Unbind();
+            this.Attenuation = attenuation;
         }
 
         /// <summary>
         /// 
         /// </summary>
         /// <returns></returns>
-        public abstract mat4 GetProjectionMatrix();
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <returns></returns>
-        public abstract mat4 GetViewMatrix();
-
-        #region ITextureSource 成员
-
-        /// <summary>
-        /// 
-        /// </summary>
-        public Texture BindingTexture
+        public override string ToString()
         {
-            get { return this.framebufferProvider.BindingTexture; }
+            return string.Format("{0}: Color:{1}, {2}", this.GetType().Name, this.diffuse, this.specular);
         }
 
-        #endregion
     }
 }
