+++ CSharpGL/Scene/Lights/SpotLight.cs
@@ -6,83 +6,34 @@ using System.Text;
 namespace CSharpGL
 {
     /// <summary>
-    /// 
+    /// Spot light.
     /// </summary>
-    public class SpotLight : LightBase
+    public sealed class SpotLight : LightBase
     {
         /// <summary>
-        /// 
+        /// Direction to the light's position.
         /// </summary>
-        public vec3 Target { get; set; }
+        public vec3 Direction { get; set; }
+
+        /// <summary>
+        /// Cos(angle).
+        /// </summary>
+        public float CutOff { get; set; }
 
         /// <summary>
         /// 
         /// </summary>
-        public float Angle { get; set; }
-
-        /// <summary>
-        /// 
-        /// </summary>
-        public float Near { get; set; }
-
-        /// <summary>
-        /// 
-        /// </summary>
-        public float Far { get; set; }
-
-        private Attenuation attenuation = new Attenuation(1.0f, 0.0f, 0.0f);
-        /// <summary>
-        /// 
-        /// </summary>
-        public Attenuation Attenuation
+        /// <param name="position">light's position.</param>
+        /// <param name="direction">Direction to the light's position.</param>
+        /// <param name="cutOff">Cos(angle)</param>
+        /// <param name="attenuation"></param>
+        public SpotLight(vec3 position, vec3 direction, float cutOff, Attenuation attenuation = null)
+            : base(attenuation == null ? new Attenuation(1.0f, 0.0f, 0.0f) : attenuation)
         {
-            get { return attenuation; }
-            set { attenuation = value; }
-        }
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <param name="position"></param>
-        /// <param name="target"></param>
-        /// <param name="angle">in degree.</param>
-        /// <param name="near"></param>
-        /// <param name="far"></param>
-        public SpotLight(vec3 position, vec3 target, float angle, float near = 0.1f, float far = 50.0f)
-        {
-            if (position == target) { throw new ArgumentException(); }
-            if ((position - target).normalize() == upVector
-                || (target - position).normalize() == upVector) { upVector = new vec3(1, 1, 1); }
-
             this.Position = position;
-            this.Target = target;
-            this.Angle = angle;
-            this.Near = near;
-            this.Far = far;
+            this.Direction = direction;
+            this.CutOff = cutOff;
         }
 
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <returns></returns>
-        public override mat4 GetProjectionMatrix()
-        {
-            const float aspectRatio = 1.0f;
-
-            mat4 projection = glm.perspective((float)(this.Angle * Math.PI / 180.0), aspectRatio, this.Near, this.Far);
-            return projection;
-        }
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <returns></returns>
-        public override mat4 GetViewMatrix()
-        {
-            mat4 view = glm.lookAt(this.Position, this.Target, this.upVector);
-            return view;
-        }
-
-        private vec3 upVector = new vec3(0, 1, 0);
     }
 }
